OpenAPIv3: description bugs
1. Include description when hashing an attribute.

This means that if you have two types in your design which are
structurally identical but with different descriptions on some
attributes they will no longer be deduped. This is useful if
descriptions are used in generated documentation or clients.

2. Include descriptions on reference attributes in the generated schema

Previously any attribute which was defined as a `ref` to another schema
in the output would not have the description included. This is now
included!